Move subquery validation out of CheckAnalysis
